色々刷新

+++ packages/juntohyo-abstract-server/src/routes/delete_election.ts
@@ -1,17 +1,25 @@
 import { Hono } from "hono";
-import { verifyToken } from "../utils/authentication";
+import { verifyToken } from "../middlewares/token";
 import { hexToBinary } from "../utils/id_convert";
 import type { Env } from "../types";
+import { verifyTurnstile } from "../middlewares/turnstile";
+import { error, ErrorCodes } from "../utils/error";
 
-export const deleteElectionRoute = new Hono<Env>().delete("/elections/:id", async (c) => {
-    const id = c.req.param("id")
-    await verifyToken(c, c.req.header("Authorization") ?? "", id);
+export const deleteElectionRoute = new Hono<Env>().delete("/elections/:id",
+    verifyTurnstile,
+    verifyToken,
+    async (c) => {
+        const id = c.req.param("id");
+        if (c.var.election !== parseInt(id)) {
+            return error(c, ErrorCodes.IncorrectRequest);
+        }
 
-    await c.env.ELECTIONS_KV.delete(id);
-    
-    const blob_id = hexToBinary(id);
+        await c.env.ELECTIONS_KV.delete(id);
+        
+        const blob_id = hexToBinary(id);
 
-    await c.env.VOTES_DB.prepare("DELETE FROM votes WHERE election = ?;").bind(blob_id).run();
+        await c.env.VOTES_DB.prepare("DELETE FROM votes WHERE election = ?;").bind(blob_id).run();
 
-    return c.body(null, 204);
-});
+        return c.body(null, 204);
+    }
+);
